+++ lib/packaging/config/params.rb
@@ -55,6 +55,8 @@ module Pkg::Params
                   foss_platforms
                   freight_archive_path
                   freight_conf
+                  gcp_signed_bucket
+                  gcp_tosign_bucket
                   gem_default_executables
                   gem_dependencies
                   gem_description
@@ -228,6 +230,8 @@ module Pkg::Params
               { :var => :foss_only,               :envvar => :FOSS_ONLY,       :type => :bool },
               { :var => :foss_platforms,          :envvar => :FOSS_PLATFORMS,  :type => :array },
               { :var => :freight_archive_path,    :envvar => :FREIGHT_ARCHIVE_PATH },
+              { :var => :gcp_signed_bucket,       :envvar => :GCP_SIGNED_BUCKET },
+              { :var => :gcp_tosign_bucket,       :envvar => :GCP_TOSIGN_BUCKET },
               { :var => :gem_host,                :envvar => :GEM_HOST },
               { :var => :gpg_key,                 :envvar => :GPG_KEY },
               { :var => :gpg_name,                :envvar => :GPG_NAME },